feat: upgrade MCP SDK to 0.8.0, add health resource and cache metadata
Upgrade ModelContextProtocol SDK from 0.4.1 to 0.8.0. Add uno://health
MCP resource, HealthReport fields (UnoSdkVersion, DiscoveryDurationMs),
cache invalidation metadata (workspace hash, SDK version), and atomic
cache writes.
