+++ src/Uno.UI.DevServer.Cli/Uno.UI.DevServer.Cli.csproj
@@ -29,8 +29,8 @@
 		<PackageReference Include="Microsoft.Extensions.Logging" Version="9.0.0" />
 		<PackageReference Include="Microsoft.Extensions.Logging.Console" Version="9.0.0" />
 		<PackageReference Include="Microsoft.Extensions.DependencyInjection" Version="9.0.0" />
-		<PackageReference Include="ModelContextProtocol" Version="0.4.1-preview.1" />
-		<PackageReference Include="ModelContextProtocol.AspNetCore" Version="0.4.1-preview.1" />
+		<PackageReference Include="ModelContextProtocol" Version="0.8.0-preview.1" />
+		<PackageReference Include="ModelContextProtocol.AspNetCore" Version="0.8.0-preview.1" />
 		<PackageReference Include="Serilog.Extensions.Logging.File" Version="3.0.0" />
 		<PackageReference Include="Spectre.Console" Version="0.49.1" />
 	</ItemGroup>
